requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
tensorflow-cpu==2.10.0
pillow==9.5.0
numpy==1.23.5
openai==0.28.1
python-multipart==0.0.6
pytest==7.4.0
requests==2.28.2